Feat security vol4

+++ src/main/java/by/tms/gradework_pm/controller/ProjController.java
@@ -7,6 +7,8 @@ import by.tms.gradework_pm.entity.Project;
 import by.tms.gradework_pm.exception.BusinessException;
 import by.tms.gradework_pm.service.EmployeeService;
 import by.tms.gradework_pm.service.ProjectService;
+import by.tms.gradework_pm.util.ProjectRoles;
+import by.tms.gradework_pm.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -28,30 +30,47 @@ public class ProjController {
     public String displayProjects(Model model) {
         List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
         model.addAttribute("projects", projects);
-        return "projects/list-projects";
+
+        if (isEquals()) {
+            return "projects/list-projects";
+        } else {
+            return "projects/list-projects-user";
+        }
     }
 
     @GetMapping("/new")
     public String displayProjectForm(Model model) {
 
-        Project project = new Project();
-        List<Employee> employees = employeeService.getAllEmployees();
-        project.setEmployees(employees);
-        model.addAttribute("project", project);
-        model.addAttribute("employees", employees);
+        if (isEquals()) {
+            Project project = new Project();
+            List<Employee> employees = employeeService.getAllEmployees();
+            project.setEmployees(employees);
+            model.addAttribute("project", project);
+            model.addAttribute("employees", employees);
 
-        return "projects/new-project";
+            return "projects/new-project";
+        } else {
+            List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
+            model.addAttribute("projects", projects);
+            return "projects/list-projects-user";
+        }
     }
 
-    @PostMapping(value = "/save")
+    @PostMapping("/save")
     public String createProject(Project project, BindingResult bindingResult, Model model) {
 
-        try {
-            projectService.saveNewProject(project);
-            return "redirect:/projects";
-        } catch (Exception e) {
-            model.addAttribute("message", e.getMessage());
-            return "errorpages/error";
+        if (isEquals()) {
+            try {
+                projectService.saveNewProject(project);
+                return "redirect:/projects";
+            } catch (Exception e) {
+                model.addAttribute("message", e.getMessage());
+                return "errorpages/error";
+            }
+        } else {
+            List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
+            model.addAttribute("projects", projects);
+            return "projects/list-projects-user";
         }
 
     }
@@ -63,13 +82,19 @@ public class ProjController {
 
         final String MESSAGE = "Error on project update";
 
-        try {
-            Project tmpProject = projectService.findById(id);
-            model.addAttribute("project", tmpProject);
-            return "projects/update-project";
-        } catch (BusinessException e) {
-            model.addAttribute("message", MESSAGE);
-            return "errorpages/error";
+        if (isEquals()) {
+            try {
+                Project tmpProject = projectService.findById(id);
+                model.addAttribute("project", tmpProject);
+                return "projects/update-project";
+            } catch (BusinessException e) {
+                model.addAttribute("message", MESSAGE);
+                return "errorpages/error";
+            }
+        } else {
+            List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
+            model.addAttribute("projects", projects);
+            return "projects/list-projects-user";
         }
     }
 
@@ -79,20 +104,32 @@ public class ProjController {
 
         final String MESSAGE = "Error on project update";
 
-        projectService.updateProject(project);
-        return "redirect:/projects";
+        if (isEquals()) {
+            projectService.updateProject(project);
+            return "redirect:/projects";
+        } else {
+            List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
+            model.addAttribute("projects", projects);
+            return "projects/list-projects-user";
+        }
 
     }
 
     @GetMapping("/delete")
     public String deleteProject(@RequestParam("id") Long id, Model model) {
 
-        try {
-            Project proj = projectService.findById(id);
-            projectService.removeProject(id);
-            return "redirect:/projects";
-        } catch (BusinessException e) {
-            return "error";
+        if (isEquals()) {
+            try {
+                Project proj = projectService.findById(id);
+                projectService.removeProject(id);
+                return "redirect:/projects";
+            } catch (BusinessException e) {
+                return "errorpages/error";
+            }
+        } else {
+            List<ProjectDtoWithStringDate> projects = projectService.getAllProjects();
+            model.addAttribute("projects", projects);
+            return "projects/list-projects-user";
         }
     }
 
@@ -107,5 +144,9 @@ public class ProjController {
 
     }
 
+    private static boolean isEquals() {
+        return SecurityUtil.getRole()
+                .equals(ProjectRoles.ROLE_ADMIN.name());
+    }
 
 }
